This changes chanid in programinfo from a string to a uint.
From the number of files touched it looks like a ridiculously large
commit, but I just wanted to to fix the params on one method when
fixing #664 and realized it would save a lot of trouble in the future
if we just changed this to a uint in programinfo instead of converting
this integer to and from a string all over the place.



git-svn-id: http://svn.mythtv.org/svn/trunk@7907 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythfrontend/manualbox.cpp
@@ -332,7 +332,7 @@ void ManualBox::startClicked(void)
     progInfo.subtitle = m_subtitle->text();
     progInfo.description = m_descString;
     progInfo.category = m_categoryString;
-    progInfo.chanid = m_chanidString;
+    progInfo.chanid = m_chanidString.toUInt();
 
     QDateTime now = QDateTime::currentDateTime();
     // Make sure we set the seconds to 0
